requirements.txt
python-telegram-bot[job-queue]==21.0
asyncpg==0.29.0
python-dotenv==1.0.0
APScheduler==3.10.4
aiohttp==3.9.1
